changed all button to shadcn components
